MUL-140: Move template to header
